team-11/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
langgraph>=0.2.40
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
mcp>=1.1.0
httpx>=0.27.0
sse-starlette>=2.1.0
